Fix class __doc__ type
