+++ Tests/test_regressions.py
@@ -824,24 +824,6 @@ class C:
         except SomeOtherError:
             pass
 
-
-    @unittest.skipIf(is_netcoreapp, 'no clr.CompileModules')
-    @skipUnlessIronPython()
-    def test_gh1357(self):
-        filename = os.path.join(self.temporary_dir, 'gh1357_%d.py' % os.getpid())
-        dll = os.path.join(self.temporary_dir, "test_%d.dll" % os.getpid())
-        with open(filename, 'w') as f:
-            f.write('{(1,): None}')
-
-        import clr
-        try:
-            clr.CompileModules(dll, filename)
-        except:
-            Fail('Failed to compile the specified file')
-        finally:
-            os.unlink(filename)
-            os.unlink(dll)
-
     @skipUnlessIronPython()
     def test_gh1435(self):
         import clr
